refactor CLI runtime orchestration into modules
